requirements.txt
# requirements.txt
pandas
torch
sentence-transformers
fastapi
uvicorn[standard]
python-multipart
numpy
scikit-learn
brave-search
google-search-results
requests
openai
python-dotenv
rank-bm25
streamlit
openpyxl
xlsxwriter
gradio
psutil